+++ src/modules/cjsToEsm.mjs
@@ -1,0 +1,38 @@
+import * as path from "path";
+import { release, version } from 'os';
+import { fileURLToPath } from 'url';
+import { createServer as createServerHttp} from 'http';
+import './files/c.js';
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
+
+const random = Math.random();
+
+let unknownObject = 'oloo';
+
+async function loadModules() {
+    if (random > 0.5) {
+        // await import('./files/a.json', { assert: { type: 'json' }}).then((res) => unknownObject = res);
+    } else {
+        // await import('./files/b.json', { assert: { type: 'json' }}).then((res) => unknownObject = res);
+    }
+}
+
+loadModules();
+
+console.log(`Release ${release()}`);
+console.log(`Version ${version()}`);
+console.log(`Path segment separator is "${path.sep}"`);
+
+console.log(`Path to current file is ${__filename}`);
+console.log(`Path to current directory is ${__dirname}`);
+
+const createMyServer = createServerHttp((_, res) => {
+    res.end('Request accepted');
+});
+
+export {
+    createMyServer,
+    unknownObject
+}
